fix: celery configuration and docker-compose
