+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 venv/
 env/
 .env
+.DS_Store
