--- apps/clnapp/tasks/__init__.py
@@ -1,1 +1,0 @@
-from .assign_badges import assign_pionneer_badge, assign_star_badge
