--- apps/clnapp/tasks/config.py
@@ -1,6 +1,0 @@
-from celery import Task
-
-
-class BaseTaskWithRetry(Task):
-    autoretry_for = (Exception, KeyError)
-    retry_kwargs = {"max_retries": 7, "countdown": 5}
