+++ apps/codeln3d/celery.py
@@ -1,0 +1,11 @@
+import os
+from django.conf import settings
+
+from celery import Celery
+
+
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "codeln3d.settings")
+
+app = Celery("codeln3d")
+app.config_from_object("django.conf:settings", namespace="CELERY")
+app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
